Endorse unsigned char vs. char for encoded UTF-8
Previously I considered Rebol's decision to use an unsigned char for REBYTE
to be just a nuisance once you turn on sign checking.  Because that's what it
was.  But I realized that actually, the type error you got on sign incorrectness
could help distinguish unencoded characters from encoded bytes, even in C.
It's a good type check and so Ren/C++ is adopting it instead of just casting it
away… when talking about bytes, use unsigned char (or REBYTE if available;
it's not available in the Red version since it doesn't use Rebol headers).

http://blog.hostilefork.com/c-casts-for-the-masses/