Remove cloned c-do.c, c-frame.c, m-gc.c
Initially Ren/C++ was prototyped against the GitHub master for rebol
and for rebolsource.  There was not a stated mission to actually fork it
and clean it for the purposes of serving as a library.  Three files had to
be borrowed to fork, and it would have taken more.  The necessary
changes are now done on the Ren/C project, and from now on you will
have to build Ren/C++ against Ren/C.  If rebolsource and/or rebol
ever take the changes back, then you will be able to build against them
again.