Adjust to Compare_Modify_Values semantics
When Rebol didn't try to make its constness ever be correct, errors about
passing const to a non-const routine were generally "crying wolf".  But with
const correctness actually making an effort to be right in Rebol, it pointed
out this *wasn't* crying wolf.  Compare_Values actually did mutate.  Given
the surprise of that, I gave it a Compare_Modify_Values name in Ren/C
and patched the mistake here so it doesn't corrupt the values you are
comparing.