Get rid of "volatile" cv-qualifier on Loadables
Although this was marked volatile, I think that was from a time when
the "const" data was passed in and the idea was that the values would
be changed in-place if they were loadables.  It doesn't work that way
now.  I'm not sure if for efficiency it will be changed to let you "upgrade"
a Loadable to its loaded value at some point.  In the meantime, it should
just be correct as a const--which is how it's being handled.