Qt version check for QtInfoMsg
Qt 5.5 added another type of message to the error handler callback,
and since we've got the "must have cases for all enum values"
warning turned on, it has to be in the list if it's there.

(The warning is a bit silly, and perhaps not worth worrying about, so
if things like this turn out to happen a lot and uglify the code too much
it might need to be turned off.  Until then, it's probably (?) helpful.)