Adapt to THROW/CATCH changes in Ren/C
Although there were some pretty significant changes to how
throw/catch and error handling are done (two different things
in Rebol!), the changes were done mindful of the impacts on
Ren/C++.  So they were not as difficult to absorb as they
might be, and generally made the code better.

It also brought out some nuance to how THROW and
CATCH wasn't being done right, so this hopefully puts it
on a path to handling those problems.

Because the GC is enabled, the problem of hooking with
that so that the GC is aware of the outstanding references
in Ren/C++ is still outstanding.  All the information is there
to do it, just a question of making a decision about how.
But until it's fixed, the tests will have problems.