Make inefficient UTF-8 molding method free series
A Ren/C change formalized the GC un-managedness of the
result of doing a UTF8 conversion into a series.  So the name of
the routine to call changed, also the series has to be freed.

Note that this is a very roundabout way to mold the data as
UTF8 and could be improved upon.