Tweak aggregate transcoded series lifetime management
The temporary aggregate series built when constructing-and-maybe
applying might wind up becoming implicitly managed by the
Val_Init_Block inside of routines like DO, so it had to be explicitly
managed if that were the case to keep the SAVE/UNSAVE pairings
consistent.

Also, due to the modifications of error handling, it does not need to
be freed in case of an error during load or evaluation.