Use new Rebol GC hook and remove threading
The very original concept of the Ren/C++ common API to Rebol and
Red would be that the C++ classes would get value cells back from the
interface layer at refcount 1, and keep them until they notified the layer
that they were no longer in use.  When values were copied they would
raise the reference count, and drop them to 0 in the destructor.

However, having such reference counting dependent on the Rebol or
Red layer means they'd still have to be holding a tracking table of the
relationship to the values held by C++.  Also, implementing such
tables is more work on the low-level side than it is to manage in C++,
where the size requirements on individual values are more relaxed as
there are fewer of them...as well as data structures being more
available.

In order to get the long-standingly necessary connection to the GC to
prevent crashing on outstanding values with series references, this
bites the bullet and works with a simple hook added to Ren/C.  That
hook calls out to a void function with no parameters, registered to do
callbacks marking REBVALs.  The tracking information for the list
of values held comes from a doubly linked list which is threaded
through the ren::Value items themselves, maintained for those value
types that require GC interaction.

This also removes the portion of the code geared toward interacting
with Rebol's thread-local storage builds...which are not currently
active and were overhead with no benefit.  (Not required for Ren
Garden to be multithreaded, for instance.)