Fix warning about not being able to call main
In some of the tinkering in order to make it possible to use Ren
Garden with MoleBox on Windows (or things like it), I'd wound up
parameterizing main...which I don't think I knew at that time was a
violation of C/C++ standards--but saw something about that since.

For some reason it didn't trigger a warning at the time it was done,
but it raised one when trying to build for Windows now.

The ultimate fate of this trickery for single-executable packing is not
known, but it's easy enough to make a proxy main for now.