Updates for Ren Garden on Windows MinGW 4.9.1
This works around what appears to be a bug in MinGW 4.9.1, which
is the packaged build that came with the QtCreator I'd installed.  It also
capitalizes the RL_START call for the RL_Api so that it does the call
through the table vs. trying to direct link it.

Adds an assert related to catching the bug.