Add App_Instance initialization for Windows
Initially thought to be something relevant only to the host-main.c file
or things-that-created-Windows, it turned out that one of those
Windows is important...an invisible one which is used for managing
a message queue.  This amends the remark that it served no point
but to get Rebol to link, and initializes it so those queues will work.