Factor evaluator and initialize Ren/C++ on its thread
It had been a plan already to break out the evaluator worker thread as
a separate piece to be used by more than just console evaluations.
For instance, if open-ended evaluations are run on the GUI thread--as
with the watchlist right now--then they could hang and there would be
no way to cancel them.

The need for this accelerated when it was realized that thread affinity
played a role in the behavior of the networking on windows.  e.g. if
Rebol services were initialized on the GUI thread first, then the worker
thread would be unable to use certain features (like http reading).

This is a somewhat rushed job to start that factoring that did the trick,
but could be more polished in terms of the effects on the startup
sequence...which now is a bit latent.  However, it gets http and https
to work on Windows.