No -fPIC on WIN32, CMP0020, RIGOROUS, sprucing
This is a combined set of patches just bringing the CMakeLists.txt up
to date after people had started trying to build on Windows with CMake
2.8.12 and the latest Qt Creator.  Since use of the -Werror flag that strictly
turns errors up had been a problem and is likely to be a blocking point for
some people in the future, this also adds `-DRIGOROUS=no` as an option
for making the build process more lax without having to edit CMakeLists.txt
directly by hand.

It also just includes improvements and updates on the notes, removal of
outdated warnings, less excessive double-and-triple-spacing, etc.  Rather
than a list of commits the notes are covered by the improved commentary
done in the CMakeLists.txt file itself.