Alias std::experimental::optional into ren::optional
Given the pending changes to UNSET!, it is apparent that the ren::Unset
value type is going to need to be replaced by some kind of optional
construct.  Although it's not yet time to bump up the dependency of Ren/C++
from C++11 to C++17, there is an available std::experimental::optional
implementation that works in C++11 (already being used in Ren Garden).

This puts it into the ren namespace and includes as ren::optional.
It doesn't do anything with it yet in Ren/C++ regarding the unsets (that will
be a bigger change).  But worth noting is that Ren Garden imports the
whole ren namespace already.  So its usage of undecorated optional now
"just works".