exception (non-Rebol) MessageBox on GUI thread
Ren Garden does a catch-all for C++ exceptions so the program
does not crash, which is one of the good things about C++ exception
handling.  It notifies with a message box and keeps running.

But these message boxes were being delivered on the GUI thread,
which meant if such exceptions happened on the worker (which was
more rare, now more common) then you'd get a Qt assert about
trying to create a UI element from the worker thread.  This queues
up the message box to the GUI.