UNSET! => optional<Value> conversion commit
This is a far-reaching (but clarifying) change.  It absorbs impacts from
changing Rebol's notion of UNSET! from being a legitimate atomic
type (e.g. something you can put into a block like any other value) to
something that represents a transitory "optional" state of a variable
not being set, and can only exist in the abstract during an evaluation.

It was performed on the binding before making the change to the
language itself.  That is interesting in the sense that it points out that
Red would not (for instance) need to make a similar prohibition on
UNSET! in a block to use Ren/C++.

Most relevant notes are in the source.  Change to clients is to use
`optional<AnyType>` anywhere an unset would be legal, such as in
processing results from arbitrary evaluation.  `nullopt` replaces `unset`.