Add README.md to include directory so it gets made
The optional migration removed optional.hpp from include/optional in the
workbench directory, so it was no longer created by git.  This made the build
script angry because it included include/.  Rather than delete the inclusion
this adds README.md to keep the directory.  (It had to go one way or the
other, and includes in the future are likely... if only to move the Ren Garden
.h files into it.)