Update WATCH dialect, evaluates blocks
This updates the WATCH dialect so that it will evaluate blocks when
are given as source literals on each Repl cycle, and paren evaluation
will run just once and be captured.  It also provides /DIALECT to
request that a watched value that comes from evaluation be treated
as a source-level instruction as if a literal had been given.  Hence
`watch (2 - 3)` will watch the numeric value -1, while watch number 1
is deleted with `watch/dialect (2 - 3)` just as if you typed `watch -1`

This also makes it so that `watch x` actually does a watch on `:x`,
which means that it's not an error to get an undefined value...also that
single arity functions will not be run.  `watch [x]` will evaluate x as an
expression which includes running functions.