AnyValue::isXXX() functions => ren::is<Type>(value)
Instead of writing foo.isBlock() for instance you now say is<Block>(foo)

This is better for a couple of reasons.  One is that these functions were all
available in the base AnyValue class, because an AnyValue could be anything.
But nothing omitted them from the derived classes.  Which means you
wouldn't get a compile-time error if you said something like `String foo;` and
then later `foo.isBlock()`.  Not the compiler can statically check for impossible
requests for type testing.

Another benefit is the global free function can just pass on disengaged
optionals by having an overload for optional<XXX>, so you don't need to do
a separate check for nullopt.