General cleanup, don't include Rebol headers in clients
For expedience and not having to mess around with extra casts to do an
opaque type, all of the Rebol implementation header files were included
for any client.  This filled up the global namespace with Rebol decls
and macros.  So this takes care of that and removes the dependency.

It also includes other formatting and organizational changes, such as
several files accidentally getting tabs in them instead of spaces.