Remove dependency on rebol-proposals, bugfix shell
This removes Ren Garden's dependency on the rebol-proposals github
repository.

The reasoning is that now with most of the "successful" proposals in Ren/C
and not building against rebol mainline (at this time), the proper channel
for getting new proposals into the system would be through Ren/C.

It also fixes a problem with the shell dialect, where it had a for-each arg [...]
with no variable.  It was trying to use it as a "forall" somehow...this must have
never actually worked, but appeared to for some reason.