Adaptations to recent Ren-C changes
The call frame structure changed, so that needed to be brought into
sync.  Also REB_END as a "full-value-sized" datatype was eliminated,
eaning that a new choice for the "alien" type used by loadable was
necessary.  TRASH was chosen because the 0 REB_xxx value is likely
to remain reserved--even in non-debug builds--for trash.