Adapt to ANY-CONTEXT! category change
When Ren-Cpp was conceived, the idea for what the word "Context" meant
was up in the air.  It was known that OBJECT! was the mechanically
most common type of key/value store, into which things could be bound,
and that the category was called ANY-OBJECT! (for things like ERROR!,
PORT!, MODULE!, OBJECT!, ...?)  So the thought was that "Context"
could mean something more relevant to binding, a kind of container
which might group up various objects (e.g. the collection of "User"
and "Lib") and then provide a way to specify an ordering on their
usage to bind.

Since that time, "Context" has been taken for the meaning of anywhere
that a word can be bound, including the "FRAME!" of a function.  This
spawned the ANY-CONTEXT! typeset, where there is no such thing as a
CONTEXT! value...it is the name of the class to which OBJECT!, ERROR!,
etc. belong (and shortly, FRAME! as well).

The original intention of letting a context be something that can be
thrown into a list of arguments to an evaluation lives on in the
type created for it, which is called "ContextWrapper".  This ties
into some other dodgy concepts like the Context "Finder".  These are
geared toward trying to make the experience of using Ren-Cpp "feel
like" working in the REPL where there is implicit binding, without
having to explicitly pass one or more contexts in.

These concepts need review, but this changes it so that Error and
Object are instances of the AnyContext abstract class.  It is thus no
longer possible to construct a "Context"... a specific class must be
chosen (which the "Any" prefix helps to suggest).