Switch to using GC layer, pointers vs. full cells
In the original conception of Ren-Cpp, it attempted to mimic the low
level programming patterns of Rebol and Red--for instance, by having
the AnyValue contain actual cell bits, which would be managed on the
stack and not require any allocation.  In part this was done as a
thought-experiment to try and prove the "C++ can be just as efficient
as C, but more type-safe" idea.  That was mixed with the idea that
the cell initializations could be abstracted across Rebol and Red via
parameterizing templates with "initialization functions".

However, mimicking the internals of the C code wound up carrying the
responsibilities of the C code--including advertising these embedded
cells to the garbage collector.  It wound up requiring hooks that were
not as palatable for the API, as opposed to offering a "handle" approach
by which cells could be allocated and freed.

By making cell allocations with tracking cheaper, Ren-C enables a
C-style API as well as a C++-style API.  So the thought-experiment of
making the cell bits live directly inside the C++ class is being
abandoned, with greater emphasis on compatibility and interaction
between the C and C++ versions.

This is the beginning of what will be likely a wave of simplification
of the internals, while keeping the user-level coding in the tests and
examples approximately the same.