Eliminate need for "RenCall" in hooks.h
In order for RenCpp's function templating tricks to work, some of the
code has to be in header files.  This code needed to be able to speak
some kind of protocol of the "call frame".

As an interim measure, the call frame object for Rebol itself was
mirrored in the header file.  As it became more complex it was clear
a simpler protocol would be needed.  This changes it to where a RenCell
sized output slot at some location is passed as a parameter, and a
contiguous array of RenCells as arguments to the function are passed
as another.

Also contains a small fix for word initialization.