Fix Error class instances
The Error constructor has an initializer list form which will go through
the Ren-Cpp block creation process, treating text as loadable code.
But it also has an ordinary constructor which lets you use a C string
literal as the error message.  This is the difference between:

    throw Error ("a string"); // a string

    throw Error {"a string"}; // two WORD!s
    throw Error {"{a string}"}; // a string

While not glaringly obvious, it's something that's learnable.