Sync project with state of Ren-C in 2017
This makes several changes to build Ren Garden, including changes to
accomodate MSVC 2017 (which required basically giving up on `and`
instead of `&&`, which probably would annoy most C++ programmers in
any case, and if you switch back and forth with C then it's not worth
it).  Accomodate DECLARE_LOCAL and Move_Value, change names of SET_XXX
to Init_Xxx, etc.

This also removes the test harness for swapping in Red instead of
Rebol, based on the idea that the way this would be done would likely
be at a C compatibility layer that does not work in the currency of
actual value cells, but opaque pointers.  RenCpp's layering on top of
"libRebol" in this way points to getting compatibility just in terms
of however compatible it is with "libRed" in the future, if it ever
becomes relevant.

The major technical hurdle to solve was the impact of streamlining
inside Rebol on Stack_Limit for different threads.  This exposed other
weaknesses of the one-thread-at-a-time approach that was done very
blindly in the early days of RenCpp, and a workaround in Ren-C via
PUSH_UNHALTABLE_TRAP seems to take care of it for now.

This does not work on turnkey building yet, given that there is a
change pending to the Rebol build system.  Getting Travis builds will
be attended to after that is done.