Further excise Red Runtime or No Runtime
In addition to being able to select Red for the RenCpp runtime, there
was a question of whether or not you could include just the data
processing part of Rebol without needing the evaluator.  This was to
try and have more of a C++ "JSON" without a Rebol interpreter.

This never existed, in the sense that the code for evaluation was
always there.  It also becomes very hard to really figure out how much
"evaluation" one would leave out and still have a useful artifact.
(What can you do with a TIME! literal, if you can't call into the
evaluator to ask for its seconds/hours/etc?)

Instead, the goal of just making Ren-C modular and light enough to not
worry about the evaluator being there seems better.