Synchronize with Ren-C specific binding, RenCell=>REBVAL
Previous to the specific binding commit, it was basically legal to
bit-copy one REBVAL cell to another.  However, one now must use
Move_Value() to migrate cells, respecting some bit patterns that
indicate cell formatting...so the opaque "RenCell" could no longer
be handled via ordinary assignment.

This is a fairly hasty incorporation of that change, which also uses
a definition of REBVAL as an opaque type which can be the actual
definition of REBVAL when building the internal code.  This eliminates
the clunky AS_C_REBVAL and AS_REBVAL macros, cleaning things up a bit.
It does mean that accessing arguments in a frame must be done through
an API, currently called RL_Arg()...because it's not possible to
increment through the arguments of a frame when the size of that
argument is not known.

Also cleaning things up a bit is pushing the try/catch surrounding a
C++-implemented FUNCTION! out of the header file and into the
implementation file.  There already was a generic dispatcher for C++
callables which had to dispatch to the templated definition in the
header, this just moves all the exception handling and return results
for Rebol into the generic part...so the "shim" only does the parameter
unpacking.