Jan-2018 sync with Ren-C + CMakeLists.txt improvements
RenCpp and Ren Garden have been only updated periodically, due to the
lack of a full Travis build of the workbench.  This brings the project
closer to that possibility by instead of ad-hoc editing the CMakeLists,
to really try and do all the options legitimately.

Upgrades Catch test framework to version 2.

Repairs right-to-left text layout demo.

Since {foo}/bar is no longer interpreted as {foo} /bar, some of the
C string multiline merging needed to be edited to put whitespace in.
This could be remedied by a variadic technique which separated the
lines with commas instead, which is probably clearer.