Fix operator[] => pick, restore helper functionality
This brings back the basics of functionality for autocomplete, shell,
and watches.  operator[] is established as a synonym for PICK-ing.

PUSH_UNHALTABLE_TRAP is used on the error-generating thread long
enough to change the stack limit to match that thread, so that the
C_STACK_OVERFLOWING() macro doesn't get a false positive.

Small changes to the helpers.

Shorten copyright messages and change publication range to 2018.